change id to long type

+++ backend/src/main/java/dev/chinh/streamingservice/data/entity/MediaNameEntity.java
@@ -16,7 +16,7 @@ public abstract class MediaNameEntity {
 
     @Setter(AccessLevel.NONE)
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected Integer id;
+    protected long id;
 
     @Column(unique = true, nullable = false)
     @JsonProperty(value = ContentMetaData.NAME)
